pyproject.toml
[build-system]
requires = ["setuptools","setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "slangpy"
version = "0.21.1"
authors = [
    {name = "Chris Cummings", email = "chriscummings@nvidia.com"},
    {name = "Benedikt Bitterli", email = "bbitterli@nvidia.com"},
    {name = "Sai Bangaru", email = "sbangaru@nvidia.com"},
    {name = "Yong He", email = "yhe@nvidia.com"}
]
description = "Easily call Slang functions and integrate with PyTorch auto diff directly from Python."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "typing_extensions",
  "numpy>=1.26.0",
  "nv-sgl==0.13.1"
]

[tool.setuptools]
packages = ["slangpy"]

[tool.pyright]
include = ["./slangpy"]
extraPaths = ["./tests"]
pythonVersion = "3.10"
typeCheckingMode = "basic"
reportUnusedImport = "error"
reportMissingSuperCall  = "error"
reportInvalidStringEscapeSequence = "error"
reportMissingParameterType = "error"
reportMissingTypeArgument = "warning"

[tool.autopep8]
in-place = true
max-line-length = 100
#ignore:
# pep8 defaults
# E731: lambda->def
ignore = "E226,E24,W50,W690,E731"

[tool.coverage.run]
branch = true
omit = ["slangpy/tests/*"]

[tool.coverage.report]
exclude_also = [
    "if.*TYPE_CHECKING.*",
    "except ImportError",
    "raise NotImplementedError"
]
ignore_errors = true

[tool.coverage.html]
directory = "htmlcov"

[tool.isort]
known_slangpycore = ["slangpy.core"]
known_slangpy = ["slangpy"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "SLANGPYCORE", "SLANGPY", "LOCALFOLDER"]